+++ src/metadata.ts
@@ -7,7 +7,7 @@ import Package from './components/Editors/Package'
 import Resource from './components/Editors/Resource'
 import Dialect from './components/Editors/Dialect'
 import Schema from './components/Editors/Schema'
-import Report from './components/Parts/Report'
+import Report from './components/Editors/Report'
 import { render } from './render'
 
 export { Package, Resource, Dialect, Schema, Report, render }
